requirements.txt
# CE Africa Valuation Platform - Python Dependencies
# Python 3.11+ required

# Core Framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Data Processing & Analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.17.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# PDF Processing (if using PDF extractor)
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Excel Processing (if needed)
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Optional: Environment variable management
python-dotenv>=1.0.0

# Optional: Testing (for development)
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Type checking (for development)
mypy>=1.5.0
types-python-dateutil>=2.8.19
